sync: finalize校验+resume隔离与可观测性+monitor驾驶舱增强

+++ p2p-cos/src/main/java/javax/net/p2p/utils/P2PCosUtils.java
@@ -239,7 +239,7 @@ public class P2PCosUtils {
         } else {//分段上传
             int count = data.length % P2PConfig.DATA_BLOCK_SIZE == 0 ? data.length / P2PConfig.DATA_BLOCK_SIZE : data.length / P2PConfig.DATA_BLOCK_SIZE + 1;
 
-            Triple<File, File, Set<Integer>> upInfo = FileUtil.getUpInfoTmp(path);
+            Triple<File, File, Set<Integer>> upInfo = FileUtil.getUpInfoTmp(-1, path);
             Set<Integer> indexes = upInfo.getRight();
             AtomicBoolean errorTag = new AtomicBoolean(false);
             try (RandomAccessFile indexFile = FileUtil.getLockedFile(upInfo.getMiddle())) {
@@ -252,7 +252,9 @@ public class P2PCosUtils {
 
                 if (execIndexes.isEmpty()) {
                     //clean temp files
-                    upInfo.getLeft().delete();
+                    if (upInfo.getLeft() != null) {
+                        upInfo.getLeft().delete();
+                    }
                     upInfo.getMiddle().delete();
                     String md5 = SecurityUtils.toMD5(data);
                     P2PWrapper p2p = P2PWrapper.build(P2PCommand.PUT_COS_FILE_SEGMENTS_COMPLETE, new FileSegmentsDataModel(0, path, data.length, md5));
@@ -313,7 +315,9 @@ public class P2PCosUtils {
             if (errorTag.get()) {
                 throw new RuntimeException("多线程上传文件执行出现异常");
             }
-            upInfo.getLeft().delete();
+            if (upInfo.getLeft() != null) {
+                upInfo.getLeft().delete();
+            }
             upInfo.getMiddle().delete();
             String md5 = SecurityUtils.toMD5(data);
             P2PWrapper p2p = P2PWrapper.build(P2PCommand.PUT_COS_FILE_SEGMENTS_COMPLETE, new FileSegmentsDataModel(0, path, data.length, md5));
