+++ p2p-transfer/src/main/java/javax/net/p2p/utils/P2PUtils.java
@@ -500,7 +500,7 @@ public class P2PUtils  implements P2PFileService{
 		} else {//分段上传
 			int count = data.length % P2PConfig.DATA_BLOCK_SIZE == 0 ? data.length / P2PConfig.DATA_BLOCK_SIZE : data.length / P2PConfig.DATA_BLOCK_SIZE + 1;
 
-			Triple<File, File, Set<Integer>> upInfo = FileUtil.getUpInfoTmp(path);
+			Triple<File, File, Set<Integer>> upInfo = FileUtil.getUpInfoTmp(storeId, path);
 			Set<Integer> indexes = upInfo.getRight();
 			AtomicBoolean errorTag = new AtomicBoolean(false);
 			try (RandomAccessFile indexFile = FileUtil.getLockedFile(upInfo.getMiddle())) {
@@ -513,7 +513,9 @@ public class P2PUtils  implements P2PFileService{
 
 				if (execIndexes.isEmpty()) {
 					//clean temp files
-					upInfo.getLeft().delete();
+					if (upInfo.getLeft() != null) {
+						upInfo.getLeft().delete();
+					}
 					upInfo.getMiddle().delete();
 					String md5 = SecurityUtils.toMD5(data);
 					P2PWrapper p2p = P2PWrapper.build(P2PCommand.PUT_FILE_SEGMENTS_COMPLETE, new FileSegmentsDataModel(storeId, path, data.length, md5));
@@ -574,7 +576,9 @@ public class P2PUtils  implements P2PFileService{
 			if (errorTag.get()) {
 				throw new RuntimeException("多线程上传文件执行出现异常");
 			}
-			upInfo.getLeft().delete();
+			if (upInfo.getLeft() != null) {
+				upInfo.getLeft().delete();
+			}
 			upInfo.getMiddle().delete();
 			String md5 = SecurityUtils.toMD5(data);
 			P2PWrapper p2p = P2PWrapper.build(P2PCommand.PUT_FILE_SEGMENTS_COMPLETE, new FileSegmentsDataModel(storeId, path, data.length, md5));
@@ -620,7 +624,7 @@ public class P2PUtils  implements P2PFileService{
 			int blockSize = P2PConfig.DATA_PUT_BLOCK_SIZE;
 			int count = (int) (length % blockSize == 0 ? length / blockSize : length / blockSize + 1);
 
-			Triple<File, File, Set<Integer>> upInfo = FileUtil.getUpInfoTmp(path);
+			Triple<File, File, Set<Integer>> upInfo = FileUtil.getUpInfoTmp(storeId, path);
 			Set<Integer> indexes = upInfo.getRight();
 			AtomicBoolean errorTag = new AtomicBoolean(false);
 			try (RandomAccessFile indexFile = FileUtil.getLockedFile(upInfo.getMiddle())) {
